[Requests] Replaces component input with redux-form-material-ui TextFields
