+++ package.json
@@ -21,7 +21,8 @@
     "react-scripts": "1.1.0",
     "redux": "^3.7.2",
     "redux-devtools-extension": "^2.13.2",
-    "redux-form": "^7.4.2"
+    "redux-form": "^7.4.2",
+    "redux-form-material-ui": "^5.0.0-beta.3"
   },
   "scripts": {
     "start": "react-scripts start",
